appBar와 NavBar는 프레임으로 고정하고 body만 동적 작동하도록 수정
